backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
mysql-connector-python==9.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
celery==5.4.0
redis==5.2.0
pandas==2.2.3
openpyxl==3.1.5
openai==1.58.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0
email-validator==2.3.0
dnspython==2.8.0

# Data visualization libraries
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
kaleido==0.2.1  # For plotly static image export

# Google AI libraries
google-generativeai==0.7.2
google-api-python-client==2.144.0
google-auth==2.34.0